Create Leaders in Array
Approach 1:
Approach:
Let’s see the basic brute force approach. The candidates for the leaders are the array elements itself.
So, we’ll start with the first element and we’ll compare it with every other element.
If we do not find any element greater than the starting element, then the starting element is one of the possible leaders. We print it.
If we found at least one element greater than the current element, then it is not a leader.
We do this for all the rest of the elements in the array and print whenever we find a leader.


Time Complexity: O(n2), because for each element we are comparing with all the elements on its right side (nested for-loop).

Space Complexity: O(1), we are not using any extra space.






Approach 2:

In the above approach, we do a fresh traversal for each candidate. If we think carefully, we want to compare the elements on the right side only. So, what if we start from the last element.
That is, we’ll try to remember the greatest element encountered so far and we’ll use that to decide whether a candidate is a leader or not.
First, we’ll start the traversal from right. Then, we move towards the left. Whenever we encounter a new element, we check with the greatest element obtained so far.
If the current element is greater than the greatest so far, then the current element is one of the leaders and we update the greatest element.
Else, we proceed with the further elements on the left. This method prints the leaders in reverse direction of their occurences. If we are concerned about the order, we can use an extra array or a string to order.

Time Complexity: O(n), we are traversing the array only once from right to left.

Space Complexity: O(1), we are not using any extra space.